updated on 11 january

+++ supermarket-mapping/src/main/java/com/wdm/serviceimpl/OrderServiceimpl.java
@@ -1,6 +1,7 @@
 package com.wdm.serviceimpl;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -8,12 +9,14 @@ import org.springframework.stereotype.Service;
 
 import com.wdm.entity.Items;
 import com.wdm.entity.Orders;
+import com.wdm.entity.UserAccount;
 import com.wdm.exception.IdNotFoundException;
 import com.wdm.exception.OrderCustomException;
 import com.wdm.exception.ProductCustomException;
 import com.wdm.model.RequestItems;
 import com.wdm.model.RequestOrder;
 import com.wdm.repository.OrderRepository;
+import com.wdm.repository.UserAccountRespository;
 import com.wdm.response.OrderResponse;
 import com.wdm.service.OrderService;
 
@@ -22,10 +25,19 @@ public class OrderServiceimpl implements OrderService {
 
 	@Autowired
 	OrderRepository OrderRepo;
+	
+	@Autowired
+	UserAccountRespository userRepo;
 
-	public Orders placeOrder(RequestOrder requestOrder) {
+	public Orders placeOrder(RequestOrder requestOrder, long userId) {
+		
+UserAccount findById = userRepo.findById(userId).orElseThrow(() -> new IdNotFoundException("Not Found"+userId));
+		
+		String getuserRoll = findById.getuserRoll();
 		Orders orders = new Orders();
-
+			 
+		if(getuserRoll.equalsIgnoreCase("admin")) {
+		 
 		orders.setTotalPrice(requestOrder.getTotalPrice());
 		orders.setOrdertime(requestOrder.getOrdertime());
 		
@@ -37,13 +49,20 @@ public class OrderServiceimpl implements OrderService {
         }
 
 		item.setQuantity(item.getQuantity() - req.getQuantity());
+		}
 		
-		return OrderRepo.save(orders);
+		return  OrderRepo.save(orders);
 	}
 
 	public void cancelOrder(long id) throws Exception {
 		try {
-			OrderRepo.deleteById(id);
+			Optional<Orders> findById = OrderRepo.findById(id);
+			if(findById.isPresent()) {
+				OrderRepo.deleteById(id);
+			}
+			else {
+				throw new IdNotFoundException("Not Found");
+			}
 			Items item = new Items();
 			RequestItems req = new RequestItems();
 			
@@ -77,10 +96,12 @@ public class OrderServiceimpl implements OrderService {
 		catch (OrderCustomException e) {
 
 			throw new OrderCustomException("Order not found for the order Id:" + orderId,
+				
 					HttpStatus.NOT_FOUND.toString(), HttpStatus.NOT_FOUND.value());
 		}
 
 		catch (Exception e) {
+			
 			throw new Exception(e);
 		}
